requirements/my.txt
alembic==1.4.2
arrow==0.15.5
bcrypt==3.1.7
binaryornot==0.4.4
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cookiecutter==1.7.2
environs==7.4.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Caching==1.8.0
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-Static-Digest==0.1.3
Flask-WTF==0.14.3
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
jinja2-time==0.2.0
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.2
mysqlclient==1.4.6
Pillow==7.1.2
poyo==0.5.0
psycopg2-binary==2.8.5
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-slugify==4.0.0
qrcode==6.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
text-unidecode==1.3
urllib3==1.25.9
Werkzeug==1.0.1
WTForms==2.3.1
xlrd==1.2.0
